Move load-object-references to a separate file
