+++ packages/core/echo/echo-db/src/core-db/core-database.ts
@@ -21,14 +21,13 @@ import {
   type ObjectDocumentLoaded,
 } from '@dxos/echo-pipeline';
 import { type SpaceDoc, type SpaceState } from '@dxos/echo-protocol';
-import { TYPE_PROPERTIES, type EchoReactiveObject } from '@dxos/echo-schema';
+import { TYPE_PROPERTIES } from '@dxos/echo-schema';
 import { compositeRuntime } from '@dxos/echo-signals/runtime';
 import { invariant } from '@dxos/invariant';
 import { type PublicKey, type SpaceId } from '@dxos/keys';
 import { log } from '@dxos/log';
 
 import { type AutomergeContext } from './automerge-context';
-import { getObjectCore } from './doc-accessor';
 import { ObjectCore } from './object-core';
 import { getInlineAndLinkChanges } from './utils';
 import { type Hypergraph } from '../hypergraph';
@@ -296,8 +295,7 @@ export class CoreDatabase {
     });
   }
 
-  remove(obj: EchoReactiveObject<any>) {
-    const core = getObjectCore(obj);
+  removeCore(core: ObjectCore) {
     invariant(this._objects.has(core.id));
     core.setDeleted(true);
   }
@@ -501,45 +499,6 @@ export const shouldObjectGoIntoFragmentedSpace = (core: ObjectCore) => {
   // NOTE: We need to store properties in the root document since space-list initialization
   //  expects it to be loaded as space become available.
   return core.getType()?.itemId !== TYPE_PROPERTIES;
-};
-
-/**
- * EXPERIMENTAL - the API is subject to change.
- * @param objOrArray - an echo object or collection of objects with references to other echo objects.
- * @param valueAccessor - selector for a reference that needs to be loaded.
- *                        if return type is an array the method exits when all entries are non-null.
- *                        otherwise the method exits when valueAccessor is not null.
- * @param timeout - loading timeout, defaults to 5s.
- */
-export const loadObjectReferences = async <
-  T extends EchoReactiveObject<any>,
-  RefType,
-  DerefType = RefType extends Array<infer U> ? Array<NonNullable<U>> : NonNullable<RefType>,
->(
-  objOrArray: T | T[],
-  valueAccessor: (obj: T) => RefType,
-  { timeout }: { timeout: number } = { timeout: 5000 },
-): Promise<T extends T[] ? Array<DerefType> : DerefType> => {
-  const objectArray = Array.isArray(objOrArray) ? objOrArray : [objOrArray];
-  const tasks = objectArray.map((obj) => {
-    const core = getObjectCore(obj);
-    const value = valueAccessor(obj);
-    if (core.database == null) {
-      return value;
-    }
-    const isLoadedPredicate = Array.isArray(value)
-      ? () => (valueAccessor(obj) as any[]).every((v) => v != null)
-      : () => valueAccessor(obj) != null;
-    if (isLoadedPredicate()) {
-      return value;
-    }
-    return asyncTimeout(
-      core.database._updateEvent.waitFor(() => isLoadedPredicate()).then(() => valueAccessor(obj)),
-      timeout,
-    );
-  });
-  const result = await Promise.all(tasks);
-  return (Array.isArray(objOrArray) ? result : result[0]) as any;
 };
 
 enum CoreDatabaseState {
