+++ packages/core/echo/echo-db/src/core-db/index.ts
@@ -5,6 +5,7 @@
 export * from './automerge-context';
 export * from './doc-accessor';
 export * from './core-database';
+export * from './load-object-references';
 export * from './object-core';
 export * from './types';
 export * from './key-path';
